Modularize communication
